Add jwt token

+++ internal/server/storage/storage.go
@@ -5,8 +5,8 @@ import "context"
 
 // Storage is an interface to store users and secrets in persistent storage.
 type Storage interface {
-	CreateUser(ctx context.Context, login, password string) (int, error)
-	FindUser(ctx context.Context, login string) (int, string, error)
+	CreateUser(ctx context.Context, login, password string) (int64, error)
+	FindUser(ctx context.Context, login string) (int64, string, error)
 
 	Ping(ctx context.Context) bool
 	Close()
